+++ packages/mel-components/src/mel-select/src/types.ts
@@ -71,4 +71,11 @@ export interface MelSelectSlots {
    * @version @mixte/mel-components@3.3.0
    */
   'option-label'?: (option: MelSelectOption) => void;
+  /**
+   * 1. 自定义渲染选项内容 ( 仅内容部分, 而非整个选项, 和 `option-label` 的作用一致 )
+   * 2. 同时在未传入 label 插槽时代替 label 插槽
+   * @param option 选项数据
+   * @version @mixte/mel-components@3.3.0
+   */
+  'all-label'?: (option: MelSelectOption) => void;
 }